[BACKLOG-292] - copying getExecutableTypes from RepositoryResource to SystemResource.  The original endpoint stays but will call the new method.
