+++ extensions/src/org/pentaho/platform/web/http/api/resources/SystemResource.java
@@ -17,13 +17,18 @@
 
 package org.pentaho.platform.web.http.api.resources;
 
+import org.apache.commons.lang.StringUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.codehaus.enunciate.modules.jersey.ExternallyManagedLifecycle;
 import org.pentaho.platform.api.engine.IAuthorizationPolicy;
 import org.pentaho.platform.api.engine.IConfiguration;
+import org.pentaho.platform.api.engine.IContentInfo;
+import org.pentaho.platform.api.engine.IPluginManager;
+import org.pentaho.platform.api.engine.IPluginOperation;
 import org.pentaho.platform.api.engine.ISystemConfig;
 import org.pentaho.platform.engine.core.system.PentahoSystem;
+import org.pentaho.platform.repository2.unified.webservices.ExecutableFileTypeDto;
 import org.pentaho.platform.security.policy.rolebased.actions.AdministerSecurityAction;
 import org.pentaho.platform.security.policy.rolebased.actions.RepositoryCreateAction;
 import org.pentaho.platform.security.policy.rolebased.actions.RepositoryReadAction;
@@ -33,9 +38,12 @@ import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.WebApplicationException;
+import javax.ws.rs.core.GenericEntity;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.TimeZone;
 
@@ -54,6 +62,7 @@ public class SystemResource extends AbstractJaxRSResource {
 
   private static final Log logger = LogFactory.getLog( FileResource.class );
   private ISystemConfig systemConfig;
+  private IPluginManager pluginManager = PentahoSystem.get( IPluginManager.class );
 
   public SystemResource() {
     this( PentahoSystem.get( ISystemConfig.class ) );
@@ -131,6 +140,46 @@ public class SystemResource extends AbstractJaxRSResource {
       }
     }
     return new TimeZoneWrapper( timeZones, TimeZone.getDefault().getID() );
+  }
+
+  /**
+   * Retrieves the list of supported content type in the platform
+   *
+   * @return list of <code> ExecutableFileTypeDto </code>
+   */
+  @Path( "/executableTypes" )
+  @GET
+  @Produces( { APPLICATION_XML, APPLICATION_JSON } )
+  public Response getExecutableTypes() {
+    ArrayList<ExecutableFileTypeDto> executableTypes = new ArrayList<ExecutableFileTypeDto>();
+    for ( String contentType : pluginManager.getContentTypes() ) {
+      IContentInfo contentInfo = pluginManager.getContentTypeInfo( contentType );
+      ExecutableFileTypeDto executableFileType = new ExecutableFileTypeDto();
+      executableFileType.setDescription( contentInfo.getDescription() );
+      executableFileType.setExtension( contentInfo.getExtension() );
+      executableFileType.setTitle( contentInfo.getTitle() );
+      executableFileType.setCanSchedule( hasOperationId( contentInfo.getOperations(), "SCHEDULE_NEW" ) );
+      executableFileType.setCanEdit( hasOperationId( contentInfo.getOperations(), "EDIT" ) );
+      executableTypes.add( executableFileType );
+    }
+
+    final GenericEntity<List<ExecutableFileTypeDto>> entity =
+      new GenericEntity<List<ExecutableFileTypeDto>>( executableTypes ) {
+      };
+    return Response.ok( entity ).build();
+  }
+
+  private boolean hasOperationId( final List<IPluginOperation> operations, final String operationId ) {
+    if ( operations != null && StringUtils.isNotBlank( operationId ) ) {
+      for ( IPluginOperation operation : operations ) {
+        if ( operation != null && StringUtils.isNotBlank( operation.getId() ) ) {
+          if ( operation.getId().equals( operationId ) && StringUtils.isNotBlank( operation.getPerspective() ) ) {
+            return true;
+          }
+        }
+      }
+    }
+    return false;
   }
 
   private boolean canAdminister() {
